third_party/WebKit/LayoutTests/inspector/timeline/timeline-network-resource-expected.txt

CONSOLE MESSAGE: line 7: Script evaluated
Tests the Timeline API instrumentation of a network resource load

Script resource loaded

ResourceSendRequest Properties:
{
    data : {
        requestId : <string>
        requestMethod : "GET"
        url : .../inspector/timeline/timeline-network-resource.js
    }
    frameId : <string>
    stackTrace : <object>
    startTime : <number>
    type : "ResourceSendRequest"
}

ResourceReceiveResponse Properties:
{
    data : {
        mimeType : <string>
        requestId : <string>
        statusCode : 0
    }
    startTime : <number>
    type : "ResourceReceiveResponse"
}

ResourceFinish Properties:
{
    data : {
        didFail : false
        networkTime : <number>
        requestId : <string>
    }
    startTime : <number>
    type : "ResourceFinish"
}
